\name{diabetesI}
\docType{data}
\alias{diabetesI}
\title{Diabetes Data with Interactions}
\description{
The data consists of 442 patients in which the response of interest is a
quantitative measure of disease progression.  The data includes 10
baseline measurements for each patient, in addition to 45 interactions
and 9 quadratic terms, for a total of 64 variables for each patient.
The outcome is \code{Y}.
}
\source{
Efron B., Hastie T., Johnstone. I and Tibshirani R. (2004).
Least angle regression (with discussion).
\emph{Ann. Statist.}, 32:407-499.
}
\examples{data(diabetesI, package = "spikeslab")}
\keyword{datasets}
