\name{leukemia}
\docType{data}
\alias{leukemia}
\title{Golub Leukemia Gene Expression Data}
\description{
Gene expression cancer data set (Golub et al.) of samples from human
acute myeloid (AML) and acute lymphoblastic leukemias (ALL).  3571
expression values on 72 individuals.
}
\source{
Golub T.R et al. (1999).
Molecular classification of cancer: class discovery and class prediction
by gene expression monitoring.
\emph{Science}, 286(5439):531-537.
}
\examples{data(leukemia, package = "spikeslab")}
\keyword{datasets}
