\name{spikeslab.news}
\alias{spikeslab.news}
\title{Show the NEWS file}
\description{
    Show the NEWS file of the \pkg{spikeslab} package.
}
\usage{
    spikeslab.news(...)
}
\arguments{
    \item{...}{Further arguments passed to or from other methods.}
}
\value{
    None.
}
\author{
    Hemant Ishwaran \email{hemant.ishwaran@gmail.com}
}
\keyword{documentation}
